fixup thread
